add my-items module

+++ webshop-backend/items/views.py
@@ -5,13 +5,15 @@ from rest_framework import status
 from django.http import Http404
 from .models import Item
 from django.core.paginator import Paginator, EmptyPage
+from rest_framework import permissions
+
 class ItemList(APIView):
     """
     List all items, or create a new item.
     """
     def get(self, request, format=None):
 
-        queryset = Item.objects.all()
+        queryset = Item.objects.filter(order=None)
         ids = request.GET.get('ids')
         if ids is not None:
             ids = ids.split(',')
@@ -45,7 +47,7 @@ class ItemDetail(APIView):
     """
     Retrieve, update or delete a item instance.
     """
-
+    permission_classes = [permissions.IsAuthenticated]
     def get_object(self, pk):
         try:
             obj = Item.objects.get(pk=pk)
@@ -61,6 +63,8 @@ class ItemDetail(APIView):
 
     def put(self, request, pk, format=None):
         item = self.get_object(pk)
+        if request.user != item.owner:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
         serializer = ItemSerializer(item, data=request.data, partial=True)
         if serializer.is_valid():
             serializer.save()
@@ -69,6 +73,31 @@ class ItemDetail(APIView):
 
     def delete(self, request, pk, format=None):
         item = self.get_object(pk)
+        if request.user.pk != item.owner.pk:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
         item.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
+
+class MyItemList(APIView):
+    """
+    List the authenticated users on sale, sold and purchased items
+    """
+    permission_classes = [permissions.IsAuthenticated]
+    def get(self, request, format=None):
+
+        queryset = Item.objects.filter()
+
+        on_sale = request.GET.get('sale')
+        if on_sale is not None:
+            print(on_sale)
+            queryset = queryset.filter(owner=request.user, order=None)
+
+        sold = request.GET.get('sold')
+        if sold is not None:
+            queryset = queryset.filter(owner=request.user).exclude(order=None)
+
+        serializer = ItemSerializer(queryset, many=True)
+        return Response(serializer.data)
+
+
